+++ spellchecking/admin.py
