+++ spellchecking/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class SpellcheckerConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'spellchecker'
+    name = 'spellchecking'
